fix audio resampler

+++ src/modules/ffmpeg/util/audio_resampler.h
@@ -16,7 +16,7 @@ class AudioResampler
     std::shared_ptr<SwrContext> ctx;
 
   public:
-    AudioResampler(int64_t sample_rate, AVSampleFormat in_sample_fmt);
+    AudioResampler(int sample_rate, AVSampleFormat in_sample_fmt);
 
     AudioResampler(const AudioResampler&)            = delete;
     AudioResampler& operator=(const AudioResampler&) = delete;
